Changing handling of suites - step 13: Created getters for the suite object and encode/decode functions for problem index, marked some deprecated functions.

+++ code-experiments/src/coco_generics.c
@@ -135,7 +135,7 @@ void coco_problem_get_initial_solution(const coco_problem_t *self, double *initi
 }
 
 /* Commented to silence the compiler */
-static long coco_problem_get_suite_dep_index(coco_problem_t *problem) {
+static size_t coco_problem_get_suite_dep_index(coco_problem_t *problem) {
   return problem->suite_dep_index;
 }
